Fix images upload
